+++ ICD.Connect.Sources.TvTuner/Streaming/StreamTvTunerDeviceSettings.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Connect.Devices;
+using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Sources.TvTuner.Streaming
+{
+	[KrangSettings("StreamTvTuner", typeof(StreamTvTunerDevice))]
+	public sealed class StreamTvTunerDeviceSettings : AbstractDeviceSettings
+	{
+	}
+}