+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - Barco Clickshare CX30 support through implementation of their V2 API
+ - IStreamTvTunerDevice for TV Tuner devices that stream from a URI
+ - StreamTvTunerDevice - implementation of IStreamTvTunerDevice
 
 ## [7.0.2] - 2020-09-24
 ### Changed